Make only possible hints visible
This patch is really a one liner that adds a check to see
if there is any input, and if there is only draw the possible
markers and nothing else. Everything added under GDK_KEY_Backspace
is to delete previous entries and probably should have been added
a while ago.