Make key commands work on non-US keyboard layouts
`Ctrl-=` and `Ctrl-+` commands did not work on non-US like keyboard
layouts because `=` and `+` are produced differently in other layouts
e.g. `Shift` is required/not required respectively. The opposite of US
like layouts.

This patch fixes the issue by not checking if `Shift` is pressed/not
pressed when it's irrelevant to the command. Instead it uses the
`keyval` from GDK directly which has been transformed depending on
keyboard layout and pressed modifiers.

Fix #371