Clarify xterm-termite fix
This is an update to #468 to complete the instructions. I tried to go looking for termite.terminfo on my system, and didn't immediately realize that the instructions meant I had to go looking in the github repo. I think this info should be worked into the manpage (and maybe termite.terminfo included in packages), next to the instruction about `__vte_prompt_command` which is necessary to make ctrl-shift-t work.

#459, #458, #416, [#364](https://github.com/thestinger/termite/issues/364#issuecomment-288251611), [#305](https://github.com/thestinger/termite/pull/305#issuecomment-183763701), #264 show that this is a common stumbling block in the way of people using termite, which is too bad because it's a slick and sweet piece of tasty software.