Consistent use of ${PREFIX} for installation path.
I have submitted an issue with more detail. Refer to that if you would like the long version. Short version is that the omission of ${PREFIX} in the install directions for config broke the installation on my FreeBSD machine. Its addition fixed it. Reject this pull request if the omission was intentional. Also, I can submit installation troubleshooting tips for FreeBSD users if you'd like.