Make clickable URLs require Control to open
This code keeps track of whether Control is pressed (updated in
key_press_cb and key_release_cb). Flag is changed in button_press_cb.

config_info is not the best place for that flag (by looking at name),
though, but it is shared between key_press_cb and button_press_cb so it
was an obvious choice.